(feature/broken)Began adding frame lookups
After adding this feature, I noticed two things:

1. Some characters have alternate versions more common in kanji
2. I can also lookup the keyword automatically

I began adding the first, and realized I'd have to change a lot of small
parts in the script, which, to me, means a poorly factored script.

I want to associate the multiple versions of characters with each other,
so as to be able to make a note type with an alternate character, and
cards that use the note type, so that Anki can keep track of the related
characters together.

I also envision an --alternates flag:
--alternates='both'
--alternates='mine'
--alternates='kanji'

For the last one, I'd have to internally maintain a list of which
characters are the preferred kanji form, but I don't think it'd be too
long or difficult to compile.

There's got to be a simple way to build up the Character- and
VocabGroups, such that the frame number and alternates are looked up
automatically, and the different note types are generated depending on
whether an alternate is present or not.