(broken)Half added picture generation
Using the command line, unless a change is made to text2png to let it
accept pictures to generate on stdin, or passed on the command line, a
temporary file would need to be created to use it.

I think it'd be cleaner to use it as a library. That would make for
tighter integration, which would make upgrading them harder, though, so
I'll want to revisit the interfaces in text2png that I'll be using, to
make them more forgiving and robust in the face of inputs and errors.